remove invitation to check in case ProfileX.sjson is missing
This is useless. We will always have a false positive for this if
someone mods the game before creating a save file (which is often the
case for zealous UW users), and even when it's not there's really not
much Hehaistos can do (as it probably is a file permissions or cloud
synchronization issue).